+++ tests/Fixtures/UninitializedProperty/ObjectWithInitializedProperty.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Tests\Fixtures\UninitializedProperty;
+
+class ObjectWithInitializedProperty
+{
+    public ValueObject $property;
+
+    public function __construct()
+    {
+        $this->property = new ValueObject('foo');
+    }
+}